Move getThemeFilterTerms
