+++ client/state/selectors/test/get-theme-filter-terms.js
@@ -6,7 +6,7 @@ import { expect } from 'chai';
 /**
  * Internal dependencies
  */
-import getThemeFilterTerms from 'state/selectors/get-theme-filter-terms';
+import { getThemeFilterTerms } from 'state/themes/selectors';
 import { state } from './fixtures/theme-filters';
 
 describe( 'getThemeFilterTerms()', () => {
